// webapp-react/src/lib/api/models/pull-request-info.ts
export interface PullRequestInfo {
    /**
     * 
     * @type {number}
     * @memberof PullRequestInfo
     */
    id: number;
    /**
     * 
     * @type {number}
     * @memberof PullRequestInfo
     */
    number: number;
    /**
     * 
     * @type {string}
     * @memberof PullRequestInfo
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof PullRequestInfo
     */
    state: PullRequestInfoStateEnum;
    /**
     * 
     * @type {boolean}
     * @memberof PullRequestInfo
     */
    isDraft: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof PullRequestInfo
     */
    isMerged: boolean;
    /**
     * 
     * @type {number}
     * @memberof PullRequestInfo
     */
    commentsCount: number;
    /**
     * 
     * @type {UserInfo}
     * @memberof PullRequestInfo
     */
    author?: UserInfo;
    /**
     * 
     * @type {Array<LabelInfo>}
     * @memberof PullRequestInfo
     */
    labels?: Array<LabelInfo>;
    /**
     * 
     * @type {Array<UserInfo>}
     * @memberof PullRequestInfo
     */
    assignees?: Array<UserInfo>;
    /**
     * 
     * @type {RepositoryInfo}
     * @memberof PullRequestInfo
     */
    repository?: RepositoryInfo;
    /**
     * 
     * @type {number}
     * @memberof PullRequestInfo
     */
    additions: number;
    /**
     * 
     * @type {number}
     * @memberof PullRequestInfo
     */
    deletions: number;
    /**
     * 
     * @type {Date}
     * @memberof PullRequestInfo
     */
    mergedAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof PullRequestInfo
     */
    closedAt?: Date;
    /**
     * 
     * @type {string}
     * @memberof PullRequestInfo
     */
    htmlUrl: string;
    /**
     * 
     * @type {Date}
     * @memberof PullRequestInfo
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof PullRequestInfo
     */
    updatedAt?: Date;
}


/**
 * @export
 */
export const PullRequestInfoStateEnum = {
    OPEN: 'OPEN',
    CLOSED: 'CLOSED',
    unknown_default_open_api: '11184809'
} as const;
export type PullRequestInfoStateEnum = typeof PullRequestInfoStateEnum[keyof typeof PullRequestInfoStateEnum];